setup(ubuntu): Install bazel only if necessary

+++ setup/ubuntu/16.04/install_prereqs.sh
@@ -23,19 +23,25 @@ zlib1g-dev
 EOF
 )
 
-# Install bazel
-BAZEL_VERSION=3.0.0
-BAZEL_DEBNAME=bazel_${BAZEL_VERSION}-linux-x86_64.deb
-BAZEL_URL=https://github.com/bazelbuild/bazel/releases/download/${BAZEL_VERSION}/${BAZEL_DEBNAME}
-BAZEL_SHA256=dfa79c10bbfa39cd778e1813a273fd3236beb495497baa046f26d393c58bdc35
-apt-get install -y --no-install-recommends wget
-wget "${BAZEL_URL}"
-if echo "${BAZEL_SHA256}  ${BAZEL_DEBNAME}" | sha256sum -c; then
-    apt-get install -y ./"${BAZEL_DEBNAME}"
-    rm "${BAZEL_DEBNAME}"
-else
-    echo "SHA256 does not match ${BAZEL_DEBNAME}:"
-    echo "    expected: ${BAZEL_SHA256}"
-    echo "    actual  : $(sha256sum "${BAZEL_DEBNAME}")"
-    exit 1
+# Install bazel if needed
+command_exists () {
+    command -v $1 >/dev/null 2>&1;
+}
+
+if ! command_exists bazel; then
+    BAZEL_VERSION=3.0.0
+    BAZEL_DEBNAME=bazel_${BAZEL_VERSION}-linux-x86_64.deb
+    BAZEL_URL=https://github.com/bazelbuild/bazel/releases/download/${BAZEL_VERSION}/${BAZEL_DEBNAME}
+    BAZEL_SHA256=dfa79c10bbfa39cd778e1813a273fd3236beb495497baa046f26d393c58bdc35
+    apt-get install -y --no-install-recommends wget
+    wget "${BAZEL_URL}"
+    if echo "${BAZEL_SHA256}  ${BAZEL_DEBNAME}" | sha256sum -c; then
+	apt-get install -y ./"${BAZEL_DEBNAME}"
+	rm "${BAZEL_DEBNAME}"
+    else
+	echo "SHA256 does not match ${BAZEL_DEBNAME}:"
+	echo "    expected: ${BAZEL_SHA256}"
+	echo "    actual  : $(sha256sum "${BAZEL_DEBNAME}")"
+	exit 1
+    fi
 fi
